feat(KFLUXBUGS-1202): refactor the funcs of ITS
* Refactor the functions of ITS and move to helpers pkg
